UX improvements
Now the user experience in the identity page will be improved! All the validation messages now describe the case of denying the access to the user better. If there are many offensive word contained in the username only the first will be displayed as problematic. Thus the communication becomes faster and more clear for the user!